+++ src/components/Inbox/components/InboxReplyMetrics.tsx
@@ -1,9 +1,14 @@
 import React, { useMemo } from 'react';
-import { useSettingsStore } from '@src/state';
+import {
+  useMentionPublished,
+  useReplyPublished,
+  useSettingsStore,
+} from '@src/state';
 import { Text, XStack } from 'tamagui';
 import ScoreIcon from '@components/Common/Icons/ScoreIcon';
-import { ArrowDown, ArrowUp } from '@tamagui/lucide-icons';
+import { ArrowDown, ArrowUp, Clock } from '@tamagui/lucide-icons';
 import { useInboxReplyVoting } from '@components/Inbox/hooks/useInboxReplyVoting';
+import { getTimeFrom } from '@helpers/time';
 
 interface IProps {
   itemId: number;
@@ -17,6 +22,10 @@ function InboxReplyMetrics({
   type,
 }: IProps): React.JSX.Element {
   const voting = useInboxReplyVoting(itemId, commentId, type);
+
+  const published =
+    type === 'reply' ? useReplyPublished(itemId) : useMentionPublished(itemId);
+  const timestamp = useMemo(() => getTimeFrom(published), [published]);
 
   const totalScore = useSettingsStore((state) => state.totalScore);
 
@@ -40,16 +49,24 @@ function InboxReplyMetrics({
 
   if (totalScore) {
     return (
-      <XStack
-        space="$1"
-        onPress={voting.scoreVote}
-        hitSlop={3}
-        alignItems="center"
-      >
-        <ScoreIcon myVote={voting.myVote} />
-        <Text fontSize="$2" color={scoreColor}>
-          {voting.score}
-        </Text>
+      <XStack space="$2">
+        <XStack
+          space="$1"
+          onPress={voting.scoreVote}
+          hitSlop={3}
+          alignItems="center"
+        >
+          <ScoreIcon myVote={voting.myVote} />
+          <Text fontSize="$2" color={scoreColor}>
+            {voting.score}
+          </Text>
+        </XStack>
+        <XStack space="$1.5" alignItems="center">
+          <Clock size={14} color="$secondary" />
+          <Text fontSize="$2" color="$secondary">
+            {timestamp}
+          </Text>
+        </XStack>
       </XStack>
     );
   } else {
@@ -75,6 +92,12 @@ function InboxReplyMetrics({
           <ArrowDown size={14} color={downvoteColor} />
           <Text fontSize="$2" color={downvoteColor}>
             {voting.downvotes}
+          </Text>
+        </XStack>
+        <XStack space="$1.5" alignItems="center">
+          <Clock size={14} color="$secondary" />
+          <Text fontSize="$2" color="$secondary">
+            {timestamp}
           </Text>
         </XStack>
       </XStack>